namespace and enum for field type
The anonymous/unnamed namespace ensures no
unintended pollution of the global namespace,
and makes clear what the external API for this file
contains.  This makes it much easier to change
the underlying implementation, as it's guaranteed
that changes internal to the namespace won't
break other source files.

Use of a scoped enum for field type enables compiler warnings.
Specifically, use of a `switch` statement with no `default` case
will generate a compiler warning/error if missing any values.

Tested to generate 100% identical JSON on Fib256.